requirements.txt
git+https://github.com/mitsuhiko/flask.git
requests
Flask-SQLAlchemy
SQLAlchemy

# celery[redis]